Removing the read of the CA certificate bundle before appending the Verisign 1024-bit certs.
Chef is supported on platforms that do not support a UTF-8 locale (such as Wind River Linux)
and the read of the file to check if the Verisign certs are already there causes Omnibus
to unwind.  Since OpenSSL doesn't care if a cert is in a bundle twice, I'm removing the
conditional check before appending the certs and just blindly append.